change the check into health_check Signed-off-by: sichen.zhao <sichen.zhao@outlook.com>
