+++ pkg/ddc/alluxio/metadata_test.go
@@ -19,6 +19,7 @@ package alluxio
 import (
 	"context"
 	"errors"
+	"reflect"
 	"testing"
 	"time"
 
@@ -26,9 +27,12 @@ import (
 	datav1alpha1 "github.com/fluid-cloudnative/fluid/api/v1alpha1"
 	"github.com/fluid-cloudnative/fluid/pkg/ddc/alluxio/operations"
 	"github.com/fluid-cloudnative/fluid/pkg/utils/fake"
+	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/klog/v2"
+	utilpointer "k8s.io/utils/pointer"
 )
 
 func TestSyncMetadata(t *testing.T) {
@@ -83,6 +87,55 @@ func TestSyncMetadata(t *testing.T) {
 	for _, datasetInput := range datasetInputs {
 		testObjs = append(testObjs, datasetInput.DeepCopy())
 	}
+
+
+	var statefulsetInputs = []appsv1.StatefulSet{
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "hbase-master",
+				Namespace: "fluid",
+			},
+			Spec: appsv1.StatefulSetSpec{
+				Replicas: utilpointer.Int32(2),
+			},
+			Status: appsv1.StatefulSetStatus{
+				Replicas:      3,
+				ReadyReplicas: 2,
+			},
+		},
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "spark-master",
+				Namespace: "fluid",
+			},
+			Spec: appsv1.StatefulSetSpec{
+				Replicas: utilpointer.Int32(2),
+			},
+			Status: appsv1.StatefulSetStatus{
+				Replicas:      3,
+				ReadyReplicas: 3,
+			},
+		},
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "hadoop-master",
+				Namespace: "fluid",
+			},
+			Spec: appsv1.StatefulSetSpec{
+				Replicas: utilpointer.Int32(2),
+			},
+			Status: appsv1.StatefulSetStatus{
+				Replicas:      3,
+				ReadyReplicas: 3,
+			},
+		},
+	}
+
+	for _, statefulset := range statefulsetInputs {
+		testObjs = append(testObjs, statefulset.DeepCopy())
+	}
+
+
 	client := fake.NewFakeClientWithScheme(testScheme, testObjs...)
 
 	engines := []AlluxioEngine{
@@ -123,6 +176,157 @@ func TestSyncMetadata(t *testing.T) {
 		t.Errorf("fail to exec function RestoreMetadataInternal")
 	}
 	wrappedUnhookQueryMetaDataInfoIntoFile()
+}
+
+func TestSyncMetadataWithoutMaster(t *testing.T) {
+	var statefulsetInputs = []appsv1.StatefulSet{
+	}
+
+	testObjs := []runtime.Object{}
+	for _, statefulset := range statefulsetInputs {
+		testObjs = append(testObjs, statefulset.DeepCopy())
+	}
+
+	var daemonSetInputs = []appsv1.DaemonSet{
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "hbase-fuse",
+				Namespace: "fluid",
+			},
+			Status: appsv1.DaemonSetStatus{
+				NumberUnavailable: 1,
+				NumberReady:       1,
+				NumberAvailable:   1,
+			},
+		},
+	}
+
+	for _, daemonSet := range daemonSetInputs {
+		testObjs = append(testObjs, daemonSet.DeepCopy())
+	}
+
+
+	var alluxioruntimeInputs = []datav1alpha1.AlluxioRuntime{
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "hbase",
+				Namespace: "fluid",
+			},
+			Status: datav1alpha1.RuntimeStatus{
+				MasterPhase: datav1alpha1.RuntimePhaseReady,
+				WorkerPhase: datav1alpha1.RuntimePhaseReady,
+				FusePhase: datav1alpha1.RuntimePhaseReady,
+			},
+		},
+	}
+	for _, alluxioruntimeInput := range alluxioruntimeInputs {
+		testObjs = append(testObjs, alluxioruntimeInput.DeepCopy())
+	}
+
+	var datasetInputs = []*datav1alpha1.Dataset{
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "hbase",
+				Namespace: "fluid",
+			},
+			Spec: datav1alpha1.DatasetSpec{},
+			Status: datav1alpha1.DatasetStatus{
+				Phase: datav1alpha1.BoundDatasetPhase,
+				HCFSStatus: &datav1alpha1.HCFSStatus{
+					Endpoint:                    "test Endpoint",
+					UnderlayerFileSystemVersion: "Underlayer HCFS Compatible Version",
+				},
+			},
+		},
+	}
+	for _, dataset := range datasetInputs {
+		testObjs = append(testObjs, dataset.DeepCopy())
+	}
+
+	client := fake.NewFakeClientWithScheme(testScheme, testObjs...)
+
+	engines := []AlluxioEngine{
+		{
+			Client:    client,
+			Log:       fake.NullLogger(),
+			namespace: "fluid",
+			name:      "hbase",
+			runtime: &datav1alpha1.AlluxioRuntime{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "hbase",
+					Namespace: "fluid",
+				},
+			},
+		},
+	}
+
+	var testCase = []struct {
+		engine              AlluxioEngine
+		expectedErrorNil    bool
+		expectedRuntimeMasterPhase datav1alpha1.RuntimePhase
+		expectedRuntimeWorkerPhase datav1alpha1.RuntimePhase
+		expectedRuntimeFusePhase datav1alpha1.RuntimePhase
+		expectedDatasetPhase datav1alpha1.DatasetPhase
+	}{
+		{
+			engine:              engines[0],
+			expectedErrorNil:    false,
+			expectedRuntimeMasterPhase: datav1alpha1.RuntimePhaseNotReady,
+			expectedRuntimeWorkerPhase: datav1alpha1.RuntimePhaseNotReady,
+			expectedRuntimeFusePhase: datav1alpha1.RuntimePhaseNotReady,
+			expectedDatasetPhase: datav1alpha1.FailedDatasetPhase,
+
+		},
+	}
+
+	for _, test := range testCase {
+		klog.Info("test")
+		err := test.engine.SyncMetadata()
+		if err != nil && test.expectedErrorNil == true ||
+			err == nil && test.expectedErrorNil == false {
+			t.Errorf("fail to exec the SyncMetadata function with err %v", err)
+			return
+		}
+
+
+		alluxioruntime, err := test.engine.getRuntime()
+		if err != nil {
+			t.Errorf("fail to get the runtime with the error %v", err)
+			return
+		}
+
+		if alluxioruntime.Status.MasterPhase != test.expectedRuntimeMasterPhase {
+			t.Errorf("fail to update the runtime master status, get %s, expect %s", alluxioruntime.Status.MasterPhase, test.expectedRuntimeMasterPhase)
+			return
+		}
+
+		if alluxioruntime.Status.WorkerPhase != test.expectedRuntimeWorkerPhase {
+			t.Errorf("fail to update the runtime worker status, get %s, expect %s", alluxioruntime.Status.WorkerPhase, test.expectedRuntimeWorkerPhase)
+			return
+		}
+
+		if alluxioruntime.Status.FusePhase != test.expectedRuntimeFusePhase {
+			t.Errorf("fail to update the runtime fuse status, get %s, expect %s", alluxioruntime.Status.FusePhase, test.expectedRuntimeFusePhase)
+			return
+		}
+
+		var dataset datav1alpha1.Dataset
+		key := types.NamespacedName{
+			Name:      alluxioruntime.Name,
+			Namespace: alluxioruntime.Namespace,
+		}
+		err = client.Get(context.TODO(), key, &dataset)
+
+		if err != nil {
+			t.Errorf("fail to get the dataset with error %v", err)
+			return
+		}
+		if !reflect.DeepEqual(dataset.Status.Phase, test.expectedDatasetPhase)  {
+			t.Errorf("fail to update the dataset status, get %s, expect %s", dataset.Status.Phase, test.expectedDatasetPhase)
+			return
+		}
+
+	}
 }
 
 func TestShouldSyncMetadata(t *testing.T) {
