+++ pkg/ddc/alluxio/replicas.go
@@ -16,10 +16,17 @@ limitations under the License.
 package alluxio
 
 import (
+	"context"
+	"fmt"
+	"reflect"
+
+	data "github.com/fluid-cloudnative/fluid/api/v1alpha1"
 	"github.com/fluid-cloudnative/fluid/pkg/ctrl"
 	fluiderrs "github.com/fluid-cloudnative/fluid/pkg/errors"
 	cruntime "github.com/fluid-cloudnative/fluid/pkg/runtime"
 	"github.com/fluid-cloudnative/fluid/pkg/utils"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/util/retry"
 )
@@ -33,6 +40,48 @@ func (e *AlluxioEngine) SyncReplicas(ctx cruntime.ReconcileRequestContext) (err 
 			if fluiderrs.IsDeprecated(err) {
 				e.Log.Info("Warning: the current runtime is created by runtime controller before v0.7.0, scale out/in are not supported. To support these features, please create a new dataset", "details", err)
 				return nil
+			}
+			if errors.IsNotFound(err) {
+				cond := utils.NewRuntimeCondition(data.RuntimeWorkersReady, "The workers are not ready.",
+					fmt.Sprintf("The statefulset %s in %s is not found, please fix it.",
+						e.getWorkerName(),
+						e.namespace), corev1.ConditionFalse)
+
+				retry.RetryOnConflict(retry.DefaultBackoff, func() error {
+
+					runtime, err := e.getRuntime()
+					if err != nil {
+						return err
+					}
+
+					runtimeToUpdate := runtime.DeepCopy()
+
+					_, oldCond := utils.GetRuntimeCondition(runtimeToUpdate.Status.Conditions, cond.Type)
+
+					if oldCond == nil || oldCond.Type != cond.Type {
+						runtimeToUpdate.Status.Conditions =
+							utils.UpdateRuntimeCondition(runtimeToUpdate.Status.Conditions,
+								cond)
+					}
+
+					runtimeToUpdate.Status.WorkerPhase = data.RuntimePhaseNotReady
+					e.Log.Error(err, "the worker are not ready")
+
+					if !reflect.DeepEqual(runtime.Status, runtimeToUpdate.Status) {
+						updateErr := e.Client.Status().Update(context.TODO(), runtimeToUpdate)
+						if updateErr != nil {
+							return updateErr
+						}
+
+						updateErr = e.UpdateDatasetStatus(data.FailedDatasetPhase)
+						if updateErr != nil {
+							e.Log.Error(updateErr, "Failed to update dataset")
+							return updateErr
+						}
+					}
+
+					return err
+				})
 			}
 			return err
 		}
