+++ pkg/ddc/base/template_engine_test.go
@@ -130,7 +130,6 @@ var _ = Describe("TemplateEngine", func() {
 		Context("When all mount points are synced", func() {
 			It("Should sync successfully", func() {
 				gomock.InOrder(
-					impl.EXPECT().CheckExistenceOfEngine().Return(nil).Times(1),
 					impl.EXPECT().SyncMetadata().Return(nil).Times(1),
 					// impl.EXPECT().CheckAndUpdateRuntimeStatus().Return(true, nil).Times(1),
 					// impl.EXPECT().UpdateCacheOfDataset().Return(nil).Times(1),
@@ -166,7 +165,6 @@ var _ = Describe("TemplateEngine", func() {
 				ufsToUpdate.AnalyzePathsDelta()
 
 				gomock.InOrder(
-					impl.EXPECT().CheckExistenceOfEngine().Return(nil).Times(1),
 					impl.EXPECT().SyncMetadata().Return(nil).Times(1),
 					// impl.EXPECT().CheckAndUpdateRuntimeStatus().Return(true, nil).Times(1),
 					// impl.EXPECT().UpdateCacheOfDataset().Return(nil).Times(1),
