+++ pkg/ddc/jindofsx/health_check.go
@@ -158,8 +158,3 @@ func (e *JindoFSxEngine) checkFuseHealthy() (err error) {
 
 	return
 }
-
-// CheckExistenceOfEngine check the engine existed
-func (e *JindoFSxEngine) CheckExistenceOfEngine() (err error) {
-	return
-}
