+++ pkg/ddc/juicefs/health_check.go
@@ -219,8 +219,3 @@ func (j *JuiceFSEngine) checkFuseHealthy() (err error) {
 	}
 	return err
 }
-
-
-func (j *JuiceFSEngine) CheckExistenceOfEngine() (err error) {
-	return
-}
